requirements.txt
python-telegram-bot>=13.1
keyboard>=0.13.5